// src/app/modules/configuration/general/models/course_view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * course_view
 */
export interface CourseView { 
    /**
     * Gets or sets CourseTypeId
     */
    courseTypeId: number;
    /**
     * Gets or sets Name
     */
    name: string;
    /**
     * Gets or sets
     */
    description: string;
     /**
     * Gets or sets
     */
    displayName: string;
     /**
     * Gets or sets
     */
    code: string;
    id: string;
}